// platform/Emscripten/js/src/types/EditorTypes.ts
export interface EditorOptions {
    touchSlop?: number;
    doubleTapTimeout?: number;
    longPressMs?: number;
    maxUndoStackSize?: number;
}

export interface TextPosition {
    line: number;
    column: number;
}

export interface TextRange {
    start: TextPosition;
    end: TextPosition;
}